pages/your_account: split builder
