+++ web/Cargo.toml
@@ -14,6 +14,7 @@ gloo-timers = "0.2"
 http = "0.2"
 js-sys = "0.3"
 material-yew = { version = "0.1", features = ["button", "dialog", "list", "formfield", "textfield"] }
+snafu = "0.6"
 wasm-bindgen-futures = "0.4"
 yew = "0.18"
